fix(buffer): honor UnBZip cap and reject >4GiB compressed inputs
Compare successful decompress size to the caller nMaxOutput, validate
file lengths as ULONGLONG before narrowing, and extend smoke coverage
for wraparound sizes. Align changed hunks with clang-format-18.

+++ Envy/PacketLengthValidate.h
@@ -199,7 +199,7 @@ inline BOOL CBufferUnBZipOutputOk(DWORD nOutput)
 	return nOutput > 0 && nOutput <= CBUFFER_UNBZIP_MAX;
 }
 
-inline BOOL CBufferUnBZipInputOk(DWORD nCompressed)
+inline BOOL CBufferUnBZipInputOk(ULONGLONG nCompressed)
 {
 	return nCompressed > 0 && nCompressed <= CBUFFER_UNBZIP_MAX;
 }
